Correct the size of the BrokerPDU's vector
(cherry picked from commit eb609f384ee363070180f0954801c85b353c4047)

+++ libs/acn/BrokerPDU.h
@@ -51,7 +51,7 @@ class BrokerPDU: public PDU {
   void PackData(ola::io::OutputStream *stream) const;
 
   static void PrependPDU(ola::io::IOStack *stack,
-                         uint32_t vector);
+                         uint16_t vector);
 
  private:
     const PDU *m_pdu;
